refactor: separate more options related logics as composable
